requirements.txt
numpy
pandas
matplotlib
scikit-learn
tensorflow
spacy
streamlit
Pillow